change naming to correspond to the phase to which we are upgrading
