+++ Cargo.toml
@@ -2,17 +2,17 @@
 resolver = "2"
 
 members = [
-    "block_party",
     "staking_party",
     "shielding_party",
-    "shielding_reward_party",
+    "shielding_rewards_party",
+    "nam_party",
 ]
 
 default-members = [
-    "block_party",
     "staking_party",
     "shielding_party",
-    "shielding_reward_party",
+    "shielding_rewards_party",
+    "nam_party",
 ]
 
 [workspace.package]
