+++ builder/parameters/shielding-rewards-party.json
@@ -11,5 +11,5 @@
     "voting_start_epoch": 9,
     "voting_end_epoch": 18,
     "activation_epoch": 25,
-    "wasm_path": "artifacts/shielded_reward_party.wasm"
+    "wasm_path": "artifacts/shielding_rewards_party.wasm"
 }
