+++ nam_party/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "shielded_reward_party"
+name = "nam_party"
 description = "WASM transaction to transition from phase 4 to phase 5 of Namada mainnet."
 authors.workspace = true
 edition.workspace = true
